docuqa-app/requirements.txt
fastapi
uvicorn
openai
python-docx
pandas
